Once again removed lists of Config.
Registries now have a "name" attribute which is the name of the
selection, as well as active which is a reference to the selected
registry item itself (short hand for registry[registry.name])

History is now 'almost' properly handled. Remaining issue is that
setting a Field to None is forgotten.