Update Config.loadX to support "root" instead of "config"
For backwards compatibility, I updated Config.load and
Config.loadFromStream to accept older config files that use
root="root" instead of the modern root="config". A deprecation
warning will be printed to stderr in that case.

To simplify this backwards compatibility code I modified Config.load
to call Config.loadFromStream. That also improves Python 3
compatibility, since Python 3 does not support execfile.

Note that it is possible for this backwards compatibility code
to be fooled if the config override file contains an error and
that message contains the text "root"; in that situation the result
will probably be a misleading error message.