Handle ChoiceField dtype being a future str
In Python 2 future "str" is not the same type as a string
literal such that the instance check fails. Overcome this by
converting a future "str" dtype to an __builtins__.str.

This may cause problems with unicode strings and may need to
be revisited to be a bit cleverer. Additionally there is
no handling of the case where a future str dtype is used and
the allowed strings are also future str.